+++ packages/components/select/src/styles.ts
@@ -1,4 +1,9 @@
-import { radiusVariant, tv, type VariantProps } from "@jamsr-ui/utils";
+import {
+  focusVisibleClasses,
+  radiusVariant,
+  tv,
+  type VariantProps,
+} from "@jamsr-ui/utils";
 
 export const test = "";
 export const selectVariant = tv({
@@ -30,6 +35,7 @@ export const selectVariant = tv({
       "ui-active:bg-content2",
       "ui-selected:bg-content2/50",
       "ui-disabled:cursor-not-allowed ui-disabled:opacity-60",
+      ...focusVisibleClasses,
     ],
   },
   variants: {