+++ packages/components/table/src/table-body.tsx
@@ -2,13 +2,14 @@ import { useUIStyle } from "@jamsr-ui/styles";
 import type { ComponentPropsWithAs, UIProps } from "@jamsr-ui/utils";
 import { cn, deepMergeProps } from "@jamsr-ui/utils";
 import { useTableContext } from "./table-context";
+import { TableEmptyState } from "./table-empty-state";
 
 export type TableBodyProps = UIProps<"tbody">;
 
 export const TableBody = <T extends React.ElementType = "tbody">(
   $props: ComponentPropsWithAs<T>,
 ) => {
-  const { tableBody:  Props = {} } = useUIStyle();
+  const { tableBody: Props = {} } = useUIStyle();
   const props = deepMergeProps(Props, $props);
 
   const {
@@ -25,6 +26,7 @@ export const TableBody = <T extends React.ElementType = "tbody">(
   return (
     <Component className={className} {...restProps}>
       {children}
+      {!children && <TableEmptyState />}
     </Component>
   );
 };