+++ packages/components/accordion/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/accordion
 
+## 0.0.87
+
+### Patch Changes
+
+- fix components
+
 ## 0.0.86
 
 ### Patch Changes