+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/input
 
+## 0.0.106
+
+### Patch Changes
+
+- fix components
+
 ## 0.0.105
 
 ### Patch Changes