+++ packages/components/select/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @jamsr-ui/select
 
+## 0.0.102
+
+### Patch Changes
+
+- fix components
+- Updated dependencies
+  - @jamsr-ui/button@0.0.83
+  - @jamsr-ui/chip@0.0.91
+
 ## 0.0.101
 
 ### Patch Changes